Moved format_qc_results to fix circular import.
